streamlined CLI
