+++ frametree/core/cli/__init__.py
@@ -1,5 +1,10 @@
-from .base import cli
-from .store import store
-from .frameset import frameset
-from .apply import apply
-from .derive import derive
+from .store import store  # noqa: F401
+from .frameset import (  # noqa: F401
+    define,
+    add_source,
+    add_sink,
+    missing_items,
+    export,
+    copy,
+)
+from .processing import derive, apply, install_license  # noqa: F401
