+++ frametree/core/cli/frameset.py
@@ -1,28 +1,22 @@
 from __future__ import annotations
 import click
 import logging
-from pathlib import Path
-from .base import cli
 from frametree.core.frameset.base import FrameSet
 from frametree.core.store import Store
 from frametree.core.axes import Axes
 from fileformats.core import DataType
 from frametree.core.serialize import ClassResolver
+from .base import cli
 
 
 logger = logging.getLogger(__name__)
 
 
-@cli.group()
-def frameset():
-    pass
-
-
-@frameset.command(
+@cli.command(
     name="define",
     help=(
-        """Define the tree structure and IDs to include in a
-dataset. Where possible, the definition file is saved inside the dataset for
+        """Define the tree structure and IDs to include in a frame set. Where
+possible, the definition file is saved inside the dataset for
 use by multiple users, if not possible it is stored in the ~/.frametree directory.
 
 ADDRESS string containing the nick-name of the store, the ID of the dataset
@@ -123,7 +117,7 @@ def define(address, hierarchy, include, exclude, axes, id_pattern):
     dataset.save(name)
 
 
-@frameset.command(
+@cli.command(
     name="add-source",
     help="""Adds a source column to a dataset. A source column
 selects comparable items along a dimension of the dataset to serve as
@@ -223,7 +217,7 @@ def add_source(
     dataset.save()
 
 
-@frameset.command(
+@cli.command(
     name="add-sink",
     help="""Adds a sink column to a dataset. A sink column
 specifies how data should be written into the dataset.
@@ -283,7 +277,7 @@ def add_sink(address, name, datatype, row_frequency, path, salience):
     dataset.save()
 
 
-@frameset.command(
+@cli.command(
     name="missing-items",
     help="""Finds the IDs of rows that are missing a valid entry for an item in
 the column.
@@ -307,7 +301,7 @@ def missing_items(address, column_names):
             click.echo(f"'{column.name}': " + ", ".join(c.row.id for c in empty_cells))
 
 
-@frameset.command(
+@cli.command(
     help="""
 Exports a dataset from one data store into another
 
@@ -374,7 +368,7 @@ def export(
     )
 
 
-@frameset.command(
+@cli.command(
     help="""
 Creates a copy of a dataset definition under a new name (so it can be modified, e.g.
 for different analysis)
@@ -389,58 +383,3 @@ NEW_NAME for the dataset
 def copy(address, new_name):
     dataset = FrameSet.load(address)
     dataset.save(new_name)
-
-
-@frameset.command(
-    name="install-license",
-    help="""Installs a license within a store (i.e. site-wide) or dataset (project-specific)
-for use in a deployment pipeline
-
-LICENSE_NAME the name of the license to upload. Must match the name of the license specified
-in the deployment specification
-
-SOURCE_FILE path to the license file to upload
-
-INSTALL_LOCATIONS a list of installation locations, which are either the "nickname" of a
-store (as saved by `arcana store add`) or the ID of a dataset in form
-<store-nickname>//<dataset-id>[@<dataset-name>], where the dataset ID
-is either the location of the root directory (for file-system based stores) or the
-project ID for managed data repositories.
-""",
-)
-@click.argument("license_name")
-@click.argument("source_file", type=click.Path(exists=True, path_type=Path))
-@click.argument("install_locations", nargs=-1)
-@click.option(
-    "--logfile",
-    default=None,
-    type=click.Path(path_type=Path),
-    help="Log output to file instead of stdout",
-)
-@click.option("--loglevel", default="info", help="The level to display logs at")
-def install_license(install_locations, license_name, source_file, logfile, loglevel):
-    logging.basicConfig(filename=logfile, level=getattr(logging, loglevel.upper()))
-
-    if isinstance(source_file, bytes):  # FIXME: This shouldn't be necessary
-        source_file = Path(source_file.decode("utf-8"))
-
-    if not install_locations:
-        install_locations = ["file_system"]
-
-    for install_loc in install_locations:
-        if "//" in install_loc:
-            dataset = FrameSet.load(install_loc)
-            store_name, _, _ = FrameSet.parse_id_str(install_loc)
-            msg = f"for '{dataset.name}' dataset on {store_name} store"
-        else:
-            store = Store.load(install_loc)
-            dataset = store.site_licenses_dataset()
-            if dataset is None:
-                raise ValueError(
-                    f"{install_loc} store doesn't support the installation of site-wide "
-                    "licenses, please specify a dataset to install it for"
-                )
-            msg = f"site-wide on {install_loc} store"
-
-        dataset.install_license(license_name, source_file)
-        logger.info("Successfully installed '%s' license %s", license_name, msg)
